+++ src/operators/operators_initialization.hpp
@@ -6,7 +6,7 @@
 #include <algorithm>
 
 using T = std::vector<std::vector<int>>;
-using L = double;
+using L = int;
 
 //Initialization Operators ----------------------------------------------------------
